+++ packages/antd-ext/components/FormItemExt/index.ts
@@ -1,5 +1,0 @@
-import Mix from './Mix';
-import BaseFormItemExt from './FormItemExt';
-
-export default Mix;
-export { BaseFormItemExt };