+++ packages/antd-ext/components/index.ts
@@ -21,7 +21,6 @@ import {
 import LayoutExt, { BaseLayoutExt } from './LayoutExt';
 import DatePickerExt from './DatePickerExt';
 import { BaseCascaderExt } from './CascaderExt';
-import { default as FormItemExt, BaseFormItemExt } from './FormItemExt';
 
 export * from 'antd';
 import zh_CN from 'antd/es/locale/zh_CN';
@@ -52,8 +51,6 @@ export {
   LayoutExt,
   DatePickerExt,
   ProInputNumber,
-  FormItemExt,
-  BaseFormItemExt,
   Theme,
 };